requirements.txt
alembic==1.9.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
